+++ src/flamenco-manager/flamenco/startup_notification.go
@@ -45,9 +45,10 @@ func (sn *StartupNotifier) Close() {
 // Keeps trying in a goroutine until the notification was succesful.
 func (sn *StartupNotifier) Go() {
 	notification := StartupNotification{
-		ManagerURL:         sn.config.OwnURL,
-		VariablesByVarname: sn.config.VariablesByVarname,
-		NumberOfWorkers:    0,
+		ManagerURL:               sn.config.OwnURL,
+		VariablesByVarname:       sn.config.VariablesByVarname,
+		PathReplacementByVarname: sn.config.PathReplacementByVarname,
+		NumberOfWorkers:          0,
 	}
 
 	url, err := sn.upstream.ResolveURL("/api/flamenco/managers/%s/startup", sn.config.ManagerID)